[1] implemented main menu in interactive mod
